+++ Changelog.md
@@ -36,3 +36,7 @@
 - Modules may now automatically generate models for provided textures.
 - Optimized some modules, reducing how often they recompute an item's model.
 - Fixed Trident and spyglass (in stack form) not being affected by CITs.
+## 2.5
+- Added the `painting_variant` module.
+- Added the `requiredEnchantments` parameter to the `enchantment` module.
+- Started looking for modules in `variants-cit`, marked `variant-cits` as deprecated.